Added few test cases

+++ SampleMavenRestProject/src/main/java/com/ust/sampleapp/controller/UserService.java
@@ -5,12 +5,15 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import com.ust.sampleapp.exception.UserNotFoundException;
 import com.ust.sampleapp.pojo.User;
 
 @Service("userService")
 public interface UserService {
 
-	public User getUser(int id) ;
+	public User getUser(Long id) ;
+	
+//	public User getUser(int id) throws UserNotFoundException;
 	
 	public List<User> getUsers();
